// appth/src/app/models/OtroSI.ts
export class OtroSiVO {
    id: number;
    numeroContrato: String;
    cargo: number;
    area: number;
    arl: number;
    fondoPension: number;
    sede: number;
    tipoContrato: number;
    estadoContrato: number;
    eps: number;
    objetivo: String;
    fechaInicio: String;
    fechaFin: String;
    salario: number;
    tallaCamisa: String;
    tallaPantalon: String;
    tallaCalzado: number;
    idDotacion: number;
    doccontrato: String;
    catOtroSi: number;
    hojaVida: number;
    generatorCons: number;
}